% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawHullConvex_PC.R
\name{DrawHullConvex_PC}
\alias{DrawHullConvex_PC}
\title{Draw convex hulls when performing PCA}
\usage{
DrawHullConvex_PC(pc, axes, zfactor, convex = T, varnames, indnames,
  biplot = T, varlab, indlab, alpha = 0.4)
}
\value{
factorial map for the individuals
}
\description{
Draw convex hulls when performing PCA
}
