% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviz_var.R
\name{dviz_var}
\alias{dviz_var}
\title{Visualize variables in PCA}
\usage{
dviz_var(pc, namesV, scale = T, axes = c(1, 2), geom = c("text",
  "arrow"), repel = T, select.var = 0, supvar = F)
}
\value{
a factorial map for the variables
}
\description{
Visualize variables in PCA
}
